refactor: move out offer/find_content/find_nodes functions from manager.rs
